add match-based project create implementation
this is quite a terse format of matching args and invoking the gitlab
builder methods for each arg matched. While terse, it is not very
readable and may not be flexible enough to add debug or other special
handling... in later commits I revert to a more verbose
if-statement-heavy implementation, which looks a little less idiomatic,
and is definitely many more lines of code, but is probably easier to
work with in the long run.

I may come back to this match-based approach if it ends up being more
idiomatic...

PS: I also removed some of the clap.rs defaults and ensured that empty
values can't be passed.