ditch match-based builder method calls for if-based
This commit is where I ditch the match-based form of invoking the gitlab
ProjectCreateBuilder methods in favour of an if-statement approach.
I also:
* also added debug to each if block.
* fix feature access levels so that they correctly follow gitlab's
  actual API requirements, rather than the loser enumeration that the
  TPGitlab lib appears to erroneously permit.
* add issues_access_level if statement, which, for some reason, I forgot
  in earlier commits

This approach is a lot more verbose, and probably not very idiomatic,
but it makes things easier to debug, and a little more flexible to work
with. I may well go back to the match-based approach to reduce LOC and
clean things up later.