add project label support to config and attach
This adds a `labels` config variable and adds the ability to read and
write it via the various config functions. The `project attach` command
is also updated to fetch the labels from the project on the server and
save this to the git-config.

This is is done in order to seen the local repo's .git/config file with
project label data so that it can be used when creating issues,
merge_requests and such when these commands are invoked interactively.